FIX: adjust the notification timeout to match the worker timeout
The workers may take a couple of seconds to finalize their work after a
test. If it's a high volume test, that consumes lots of resources, this
shutdown period may take a while. The workers calculate this time
dynamically based on the number of connections.

Therefore, adjust the client wait time to match the time used for the
workers shutdown.