REF: implement the HDR plotting code internally
Move all the HDR-handling code that was in an external project and move
it to a module within Maestro.

Rework that code to get rid of the sloppy
HDR log processor made on top of the main class of the library the
project relies on.

Adjust the code to generate the images in parallel.

Upgrade the code to use a newer version of the Xchart library